[tests] Make test harness portable.
Prior to this change, on non-Linux, the test harness will already
attempt to use dispatch functionality when testing dispatch itself. This
does not seem intentional, and instead appears to be intended as the Mac
OS case where the platform already has Dispatch and is using that when
testing the just-built version.

Instead, make the Linux case the general `__unix__` case.