[event] Workqueue monitoring for OpenBSD.
Simply speaking, the global concurrent queue has a thread per cpu and
schedules functions or blocks on one of these threads. If these threads
are all busy, without workqueue monitoring, we cannot schedule a new
function or block until one of those threads returns. If thread(s) have
not returned however because they are asleep (say), then we could
overcommit the global concurrent queue by scheduling more work.

The Linux implementation checks each registered thread ID to see if the
corresponding thread is in a runnable state according to procfs every
second.

On some platforms without procfs, this same information is obtainable
essentially via the relevant sysctls. Libraries may exist to facilitate
issuing the correct sysctls (c.f. `kvm_getprocs`) but these were not
used here so to avoid complications with the right invocations when
linking.

A unit test is also added to test the overcommit functionality. In this
test, we tie up all cpus to prevent the scheduled threads from returning
and then attempt to run one more task. If the overcommit monitoring is
working as intended, then Dispatch will detect all the threads are
simply asleep and will attempt to overcommit. If not, the test will
notice the function we attempted to overcommit did not run and fail the
test.