Use some disposable heap arrays instead of stack.
Using a 512K buffer in this test may mean that the test process may hit
the stack size rlimit quickly and segfault. Instead of trying to reduce
the buffer sizes and then subsequently try and unwind other weird
behavior in this test, just use heap-allocated buffers. We don't actually
seem to care about the contents in this test, so we can just throw them
away immediately after we read into them.