dispatch: introduce `dispatch_STATIC` and honour `dispatch_EXPORT`
When building with CMake, `dispatch_EXPORT` is automatically defined
when building a shared library. We should additionally rely on
`dispatch_STATIC` to indicate that we are building a static library.
When building statically, we should ensure that the ABI of dispatch does
not participate in dynamic linking as otherwise any consuming library
would vend the dispatch ABI. Particularly for Linux, this would be
problematic as the symbol resolution is ill-defined and can end up with
multiple incompatible definitions.

Co-authored-by: Evan Wilde <etceterawilde@gmail.com>