build: always build libdispatch in shared mode
As per the discussion with @rokhinip, @mikeash, and @compnerd, always
build libdispatch in shared mode. The rationale for this is the
following:

  If someone does a `-static-stdlib` build of Swift code as a plugin
  which is hosted in a C/C++ environment, you cannot have two dispatch
  implementations to run this. You need a single one, and because the
  host environment may be dynamically linked - you need to dynamically
  link libdispatch (the C portions).

This prepares us to be able to build the rest of the dispatch code
(swiftDispatch) both statically and dynamically. Until we have a
resolution on BlocksRuntime, build the library both ways. Note that in
order to use a static copy of BlocksRuntime, `-static-libclosure` must
be used at *compile* time.