Install nested Swift modules
Migrating to using the nested swiftmodule structure. This avoids needing
to compute the architecture in the CMake script, using the compiler to
tell us what the triple should be, making the build more robust. This
also means that we can drop the architecture-specific subdirectory and
install the dispatch modules alongside the swiftmodules for the other
projects.