Include SwiftSupport when needed
Support for Swift is only needed when we have Swift enabled. We don't
need to include the module when it isn't available. With the change to
query the Swift compiler for the module triple, we need the Swift
language to be enabled to set the swift module triple.

Since the include was pulled from the main CMakeLists file, it's prudent
to set the include guard to prevent accidentally loading the module
multiple times.