event: ensure acquire release semantics for muxnote disposal
While it is sufficient to use the relaxed ordering for the acquire
semantics in the `_dispatch_muxnote_retain`, we need to use the acquire
release semantics on the release in `_dispatch_muxnote_release` to
ensure that any pending retains are not interrupted. This should
hopefully alleviate the occasional crashes that have been observed with
the muxnote reference counting.

Fixes: #887, #844